// apps/web/app/api/generated/schemas/getProjectsProjectIdTasks200DataItem.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Catalyst API
 * Type-safe REST API for the Catalyst project management system
 * OpenAPI spec version: 1.0.0
 */
import type { GetProjectsProjectIdTasks200DataItemDescription } from './getProjectsProjectIdTasks200DataItemDescription';
import type { GetProjectsProjectIdTasks200DataItemStatus } from './getProjectsProjectIdTasks200DataItemStatus';
import type { GetProjectsProjectIdTasks200DataItemPriority } from './getProjectsProjectIdTasks200DataItemPriority';
import type { GetProjectsProjectIdTasks200DataItemAssigneeId } from './getProjectsProjectIdTasks200DataItemAssigneeId';
import type { GetProjectsProjectIdTasks200DataItemDueDate } from './getProjectsProjectIdTasks200DataItemDueDate';

export type GetProjectsProjectIdTasks200DataItem = {
  id: string;
  projectId: string;
  /**
   * @minLength 1
   * @maxLength 200
   */
  title: string;
  description?: GetProjectsProjectIdTasks200DataItemDescription;
  status: GetProjectsProjectIdTasks200DataItemStatus;
  priority: GetProjectsProjectIdTasks200DataItemPriority;
  assigneeId?: GetProjectsProjectIdTasks200DataItemAssigneeId;
  dueDate?: GetProjectsProjectIdTasks200DataItemDueDate;
  createdAt: string;
  updatedAt: string;
};