// apps/web/app/api/generated/schemas/postProjectsProjectIdTasksBodyPriority.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Catalyst API
 * Type-safe REST API for the Catalyst project management system
 * OpenAPI spec version: 1.0.0
 */

export type PostProjectsProjectIdTasksBodyPriority = typeof PostProjectsProjectIdTasksBodyPriority[keyof typeof PostProjectsProjectIdTasksBodyPriority];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const PostProjectsProjectIdTasksBodyPriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;