// apps/web/app/api/generated/schemas/postProjectsProjectIdTasksBodyStatus.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Catalyst API
 * Type-safe REST API for the Catalyst project management system
 * OpenAPI spec version: 1.0.0
 */

export type PostProjectsProjectIdTasksBodyStatus = typeof PostProjectsProjectIdTasksBodyStatus[keyof typeof PostProjectsProjectIdTasksBodyStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const PostProjectsProjectIdTasksBodyStatus = {
  todo: 'todo',
  in_progress: 'in_progress',
  done: 'done',
} as const;